+++ app/build.gradle
@@ -24,10 +24,9 @@ android {
             storePassword 't4f5X0BSd6z1uz2ZnJXa'
         }
     }
-
     compileSdkVersion 26
     defaultConfig {
-        applicationId "by.equestriadev.superapp1"
+        applicationId 'by.equestriadev.nikishin_rostislav'
         minSdkVersion 22
         targetSdkVersion 26
         versionCode 2
@@ -41,6 +40,8 @@ android {
         ci {
             signingConfig signingConfigs.CI_Build
         }
+    }
+    productFlavors {
     }
 }
