+++ app/src/main/java/by/equestriadev/nikishin_rostislav/MainActivity.java
@@ -1,13 +1,10 @@
-package by.equestriadev.superapp1;
+package by.equestriadev.nikishin_rostislav;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 
 import com.crashlytics.android.Crashlytics;
