+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="by.equestriadev.superapp1.MainActivity">
+    tools:context="by.equestriadev.nikishin_rostislav.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
