+++ app/src/main/res/menu/menu_main.xml
@@ -1,5 +1,5 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="by.equestriadev.superapp1.MainActivity">
+    tools:context="by.equestriadev.nikishin_rostislav.MainActivity">
 </menu>
