+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Super App #1</string>
+    <string name="app_name">Rostislav Nikishin\'s app</string>
     <string name="action_settings">Settings</string>
     <string name="github">Github: Bronydell</string>
     <string name="author">Rostislav Nikishin</string>
